RFC 022: Tendermint Command Line Tools
The stock tendermint command-line tool has a large collection of subcommands.
Some are essential to the operation of the consensus node, others are ancillary
utilities.

Unlike the core operational subcommands, many of the utilities are ad hoc and
intended to address specific operational corner cases. For this reason, the
utility subcommands also have less clearly-defined behavior and are less
well-tested.  This contributes to surprising and sometimes negative user
experiences for node operators.

This RFC summarizes the existing subcommands, discusses some of the known
issues, and advances some strategies to mitigate those issues.